@@ -8,11 +8,8 @@
   "dependencies": {
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
-    "@material-ui/core": "^4.9.9",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/styles": "^4.11.3",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.1",
+    "@mui/icons-material": "^5.6.0",
+    "@mui/material": "^5.6.0",
     "@mui/x-data-grid": "^5.6.1",
     "@stardust-ui/docs-components": "^0.40.0",
     "@testing-library/jest-dom": "^5.16.2",
@@ -36,9 +33,9 @@
     "jest-dom": "^4.0.0",
     "minimist": "^1.2.3",
     "prettier": "2.5.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-async-hook": "^4.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18",
     "react-force-graph": "^1.32.1",
     "react-router-dom": "^6.2.2",
     "react-scripts": "^5.0.0",
