@@ -1,4 +1,4 @@
-import { makeStyles } from "@material-ui/core/styles";
+import { makeStyles } from "@mui/material/styles";
 import theme from "../../../../../theme";
 
 export const nodeTableStyles = makeStyles({
