@@ -1,16 +1,16 @@
 import React from "react";
 
-import Popover from "@material-ui/core/Popover";
-import Card from "@material-ui/core/Card";
-import CardActionArea from "@material-ui/core/CardActionArea";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import Button from "@material-ui/core/Button";
-import Typography from "@material-ui/core/Typography";
+import Popover from "@mui/material/Popover";
+import Card from "@mui/material/Card";
+import CardActionArea from "@mui/material/CardActionArea";
+import CardActions from "@mui/material/CardActions";
+import CardContent from "@mui/material/CardContent";
+import Button from "@mui/material/Button";
+import Typography from "@mui/material/Typography";
 
 import { notificationsStyles } from "./styles";
 
-const useStyles = notificationsStyles;
+const useStyles: any = notificationsStyles;
 
 export default function LoginNotification() {
     const classes = useStyles();
