@@ -1,6 +1,6 @@
-import { createStyles, makeStyles } from "@material-ui/core/styles";
+import { createStyles, makeStyles } from "@mui/material/styles";
 
-export const useStyles = makeStyles((theme) =>
+export const useStyles: any = makeStyles((theme: any) =>
     createStyles({
         root: {
             display: "flex",
