@@ -1,23 +1,23 @@
 import React from "react";
-import ReactDOM from "react-dom";
-
 import "./index.css";
 
-import { createTheme, ThemeProvider } from "@material-ui/core/styles";
+import {ThemeProvider } from "@mui/material/styles";
 
 import App from "./App";
 import theme from "./theme";
 import * as serviceWorker from "./serviceWorker";
+import {createRoot} from "react-dom/client";
 
-const rootElement = document.getElementById("root");
+const container = document.getElementById("root");
+const root =
+    createRoot(container);
 
-ReactDOM.render(
+root.render(
     <React.StrictMode>
         <ThemeProvider theme={theme}>
             <App />
         </ThemeProvider>
-    </React.StrictMode>,
-    rootElement
+    </React.StrictMode>
 );
 
 serviceWorker.unregister();
